+++ README.md
@@ -1,17 +1,19 @@
 # Final Fantasy Card battle
 
-> This card game uses React/Redux
-> Coming Soon Firebase
+> This card game uses React to render the game and Redux to store current game stats.
+> Bootstrap was added in to help position components
 
 ### Game Description
 
 > Card game based on FinalFantasy game art.
->
+> Players will have the ability to pick from classic heros/heroines from the FinalFantasy series
+> Test your luck against your friends in this card battle brawl.
 
 ### How To Play
 
 > 1. Pick Character Avatar
->    - `each Character has special perks to help in batle`
+>    - `each Character has different stats in health and defense`
+>    - `perks to be added in the future`
 > 2. Choose between 9 cards
 >    - `there are two attack cards used to damage opponent`
 >    - `picking bahamut damages both players`
@@ -19,4 +21,4 @@
 >    - `heal will heal current player`
 >    - `2x and 3x strike are equip cards for damage multipliers`
 
-> **NOTE:** *Coming Soon*
+> **NOTE:** *in the future, I'm insterested in adding in more cards and character abilities*
