+++ src/reducers/player_reducer.js
@@ -1,18 +1,29 @@
 import types from '../actions/types';
 import cardData from '../components/gameEngine/cardData';
+import character from '../components/gameEngine/playerStats.js';
 
-const PlayerBASE = {
-        health: 10,
-        attack: 1,
-        baseAtk:1,
-        equip: null,
-        hurt: false
+//need to make sure character stats update playerBase
 
+const Player1BASE = {
+    health: 10,
+    attack: 1,
+    baseAtk:1,
+    equip: null,
+    hurt: false
 }
+
+const Player2BASE = {
+    health: 10,
+    attack: 1,
+    baseAtk:1,
+    equip: null,
+    hurt: false
+}
+
 const DEFAULT_STATE = {
-    player1:{...PlayerBASE,
+    player1:{...Player1BASE,
     },
-    player2:{...PlayerBASE,
+    player2:{...Player2BASE,
     },
     p1Character:'',
     p2Character:''
@@ -32,10 +43,10 @@ function playerReducer( state = DEFAULT_STATE, action){
             return newState;
         case types.RESETPLAYERS:
             return {...state,
-                player1:{...PlayerBASE
+                player1:{...Player1BASE
             },
             player2:{
-                ...PlayerBASE
+                ...Player2BASE
             }
         };
         case types.DAMAGE:
@@ -44,11 +55,22 @@ function playerReducer( state = DEFAULT_STATE, action){
             newState.player2.hurt = false;
             return newState;
         case types.SETCHAR:
-            return {...state,
+            Player1BASE.health =  character[action.payload.player1].health;
+            Player1BASE.baseAtk = character[action.payload.player1].attack;
+            Player1BASE.attack = character[action.payload.player1].attack;
+            Player2BASE.health =  character[action.payload.player2].health;
+            Player2BASE.baseAtk = character[action.payload.player2].attack;
+            Player2BASE.attack = character[action.payload.player2].attack;
+            return {
+                player1:{...Player1BASE,
+                },
+                player2:{...Player2BASE,
+                },
                 p1Character: action.payload.player1,
                 p2Character: action.payload.player2
             };
         default:
+        
             return state;
     }
 }
